design_engineer_coding_challenge_v6/requirements.txt
appdirs==1.4.4
attrs==21.4.0
biopython==1.79
black==22.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
datacache==1.1.5
flake8==4.0.1
gtfparse==1.2.1
idna==3.3
iniconfig==1.1.1
mccabe==0.6.1
memoized-property==1.0.3
mock==4.0.3
mypy-extensions==0.4.3
numpy==1.22.3
packaging==21.3
pandas==1.4.1
pathspec==0.9.0
platformdirs==2.5.1
pluggy==1.0.0
progressbar33==2.4
py==1.11.0
pycodestyle==2.8.0
pyensembl==1.9.4
pyflakes==2.4.0
pyparsing==3.0.7
pysam==0.18.0
pytest==7.1.1
python-dateutil==2.8.2
pytz==2022.1
PyVCF==0.6.8
requests==2.27.1
sercol==0.1.4
serializable==0.2.1
simplejson==3.17.6
six==1.16.0
tinytimer==0.0.0
tomli==2.0.1
typechecks==0.1.0
typing-extensions==4.1.1
urllib3==1.26.9
varcode==1.0.3
